New denylist for october 17: (#8372)
The primary change was to implement 'carryover' from the previous
denylist. This means that any nodes/edges that did not have any data for
them over the 2 week sampling period are carried forward. This prevents
hotspots from simply being turned off for two weeks to escape the
denylist and then being turned back on.

All other denylist parameters remain the same as the previous week, and
no new classifiers are added.